requirements.txt
amqp==5.3.1
annotated-types==0.7.0
ansicon==1.89.0
anyio==4.10.0
asgiref==3.8.1
awsebcli==3.25.3
bcrypt==5.0.0
beautifulsoup4==4.13.4
billiard==4.2.4
blessed==1.25.0
boto3==1.35.74
botocore==1.35.99
bs4==0.0.2
cachetools==5.5.2
celery==5.6.0
cement==2.10.14
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
clr_loader==0.2.9
colorama==0.4.6
cron_descriptor==2.0.6
cryptography==45.0.6
decorator==5.2.1
Deprecated==1.3.1
distro==1.9.0
Django==5.2
dj-database-url==2.2.0
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-storages==1.14.4
django-timezone-field==7.2.1
djangorestframework==3.16.0
ecdsa==0.19.1
exceptiongroup==1.3.1
fabric==3.2.2
gevent==25.9.1
google==3.0.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
greenlet==3.3.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
invoke==2.2.1
jinxed==1.3.0
jiter==0.10.0
jmespath==1.0.1
kombu==5.6.1
numpy==2.3.2
oauthlib==3.3.1
openai==1.101.0
packaging==24.2
pandas==2.3.2
paramiko==4.0.0
pathspec==0.12.1
prompt_toolkit==3.0.52
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
PyNaCl==1.6.1
pypiwin32==223; platform_system == "Windows"
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.3.0
pythonnet==3.0.5
pytz==2025.2
pywin32==311; platform_system == "Windows"
PyYAML==6.0.3
redis==5.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.10.4
semantic-version==2.10.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
termcolor==2.5.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.8.2
wrapt==2.0.1
zope.event==6.1
zope.interface==8.1.1
